% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vortexRdata_pkg_data.R
\name{campbell}
\alias{campbell}
\title{Raw Vortex outputs from Campbell et al (2016)}
\format{Several .dat and .stdat files.}
\source{
Campbell et al. (2016). Assessing the economic benefits of starling
 detection and control to Western Australia. Australasian Journal of
 Environmental Management, 23, 81-99
 \href{https://dx.doi.org/10.1080/14486563.2015.1028486}{DOI:10.1080/14486563.2015.1028486}
}
\description{
A folder with Vortex outputs from Campbell et al (2016).
}
\examples{
campbell_dir <- system.file("extdata", "campbell", package="vortexRdata")
cat("campbell example files:\\n"); dir(campbell_dir)
}
