% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vortexRdata_pkg_data.R
\docType{package}
\name{vortexRdata}
\alias{vortexRdata}
\alias{vortexRdata-package}
\title{vortexRdata: Example data for vortexR}
\description{
\code{vortexRdata} provides real-world example data for \code{vortexR} in both
raw (Vortex output) and binary (R objects) form.
\code{vortexR} uses the raw data provided here to illustrate its capability
to parse raw Vortex output files into one R object.
}
\details{
\code{vortexR} facilitates Post Vortex Simulation Analysis (PVSA) by offering
tools to collate multiple Vortex (v10) output files into one R object, generate
plots and conduct basic analysis (e.g. pairwise comparisons of scenarios) and
more advanced statistics such as fitting of a Generalised Linear Model (GLM)
to investigate the main and the interaction effects of the variables of
interest.

\code{vortexR} has a number of functions that are useful during the
development of a Vortex project and for its analysis after completion.

\code{vortexR} facilitates the creation of plots and
computation of basic statistics to inspect the effect of changes carried out
in the Vortex project. Once the Vortex project development is completed, the same
framework used in \code{vortexR} during the development of the project can be
refined and extended to include more advanced statistical analyses or can be
easily included in Markdown documents for the creation of reports (by
converting them into pdf) or published as web-pages.

The use of \code{vortexR} ensures reproducibility and standardises analytical
approaches in population viability analysis.
}
\section{Documentations}{

Use \code{help(package = "vortexR")} for a list of \code{vortexR} functions
and their specific documentations.

A more detailed description of the package and functions can be opened with:
\code{vignette(package="vortexR", topic="User-guide")}.

More vignettes may be come available in the future. Use
\code{vignette(package="vortexR")} to see all the available vignettes.
}

\section{Citation}{

If you use \code{vortexR}, please cite:
Pacioni, C., and Mayer, F. (2017). vortexR: an R package for post Vortex
simulation analysis. Methods in Ecology and Evolution.
}

\section{Get in touch}{

Please don't hesitate to report any issues at \url{https://github.com/carlopacioni/vortexR/issues}.
Contributions to both vortexR and vortexRdata via pull requests are most welcome.
}

